Readme.md
Overview about Master Mind game :
Mastermind is a board game with an interesting history (or rather a legend?). Some game books report that it was invented in 1971 by Mordecai Meirowitz, an Israeli postmaster and telecommunications expert. After many rejections by leading toy companies, the rights were obtained by a small British firm, Invicta Plastics Ltd. The firm originally manufactured the game itself, though it has since licensed its manufacture to Hasbro in most of the world. However, Mastermind is just a clever readaptation of an old similar game called 'Bulls and cows' in English, and 'Numerello' in Italian... Actually, the old British game 'Bulls and cows' was somewhat different from the commercial version. It was played on paper, not on a board... Over 50 million copies later, Mastermind is still marketed today!

you can check a visual version of this game by this link :
http://www.archimedes-lab.org/mastermind.html

My first version of this game includes a very simple Design, where you should play by prompt message box :
because my purpose was only to be more comfortable with the OOP (Object Oreanted programming) with Typescript, next version will be with prettier design .